requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
Pillow==10.0.0
py360convert==0.1.0
pyvirtualcam==0.10.2
scipy==1.11.2
Werkzeug==2.3.7
